Revert "refactor: Message banner incident/maintenance message displays name"
This reverts commit 7359dbed29bd8db083e5b91ef8e3abc5c54d8288.
