+++ packages/website/components/messagebanner/messagebanner.scss
@@ -68,11 +68,6 @@
   &.mb-reduced-fontsize {
     @include fontSize_Tiny;
   }
-
-  &.disabled {
-    pointer-events: none;
-    cursor: default;
-  }
 }
 
 .message-banner-close-button {
